+++ lib/extension_flutter_tools.dart
@@ -5,3 +5,6 @@ export 'src/confirm_wrapper.dart';
 export 'src/date_time_button.dart';
 export 'src/editable_label.dart';
 export 'src/network_or_file_image.dart';
+export 'src/error_view.dart';
+
+export 'test/test_tools.dart';