+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+- Added an `ErrorView` component to display error messages.
+- Added `TestTools` class with some static functions useful for unit tests for Flutter apps.
+
 ## 1.0.0
 
 - Initial public version.